Added Lora send for testing

+++ Core/Inc/lora.h
@@ -145,7 +145,7 @@ LoRa_Status Lora_Set_Preamble(uint16_t preamble);
  * @param data [uint8_t*] Data buffer to send
  * @param len [uint8_t] Length of data
  */
-LoRa_Status Lora_Transmit(uint8_t* data, uint8_t len);
+LoRa_Status Lora_Transmit(uint8_t* data, size_t len);
 
 /**
  * @brief Read data from LoRa Connection
